Managed to assign last round after the waitroom! still gotta test though...
