+++ crosstalk/templates/crosstalk/SetLastRound.html
@@ -1,0 +1,21 @@
+{% extends "global/Page.html" %}
+{% load otree static %}
+
+
+{% block title %}
+
+{% endblock %}
+
+{% block styles %}
+    <style>
+        .otree-timer {
+            display: none;
+        }
+    </style>
+
+{% block content %}
+{% endblock %}
+
+<!---This page is useless. I only need it to set the subgroups/treatments in control and the last round.--->
+
+{% endblock %}
