+++ introduction_cross/models.py
@@ -21,9 +21,9 @@ class Constants(BaseConstants):
     players_per_group = 4
     num_rounds = 1
 
-    """variables for randomish next round"""
-    min_rounds = 2
-    proba_next_round = 0.5
+    # """variables for randomish next round"""
+    # min_rounds = 2
+    # proba_next_round = 0.5
 
     """
     Donation game payoff
@@ -48,32 +48,34 @@ class Subsession(BaseSubsession):
         This is for the 50% chance of another round. We create a function for clarity below in the creating_session().
         We create a list of different number of rounds that is as long as there are groups.
     """
-    def get_random_number_of_rounds(self):
-        num_groups = int(self.session.num_participants / 2)
-        list_num_rounds = []
-        for _ in range(num_groups):
-            number = Constants.min_rounds
-            while Constants.proba_next_round < random.random():
-                number += 1
-            list_num_rounds.append(number)
-        return list_num_rounds
-
-    def creating_session(self):
-        """ random last round code. With the function from above,
-                we attribute the different elements in the list to each group."""
-        list_num_rounds = self.get_random_number_of_rounds()
-        group_number_of_rounds = itertools.cycle(list_num_rounds)
-        for g in self.get_groups():
-            g.last_round = next(group_number_of_rounds)
-            print('New number of rounds', g.last_round)
-        for p in self.get_players():
-            p.participant.vars['last_round'] = p.group.last_round
-            print('vars last_round is', p.participant.vars['last_round'])
+    # def get_random_number_of_rounds(self):
+    #     num_groups = int(self.session.num_participants / 2)
+    #     list_num_rounds = []
+    #     for _ in range(num_groups):
+    #         number = Constants.min_rounds
+    #         while Constants.proba_next_round < random.random():
+    #             number += 1
+    #         list_num_rounds.append(number)
+    #     return list_num_rounds
+    #
+    # def creating_session(self):
+    #     """ random last round code. With the function from above,
+    #             we attribute the different elements in the list to each group."""
+    #     list_num_rounds = self.get_random_number_of_rounds()
+    #     group_number_of_rounds = itertools.cycle(list_num_rounds)
+    #     for g in self.get_groups():
+    #         g.last_round = next(group_number_of_rounds)
+    #         print('New number of rounds', g.last_round)
+    #     for p in self.get_players():
+    #         p.participant.vars['last_round'] = p.group.last_round
+    #         print('vars last_round is', p.participant.vars['last_round'])
+    pass
 
 
 class Group(BaseGroup):
-    """Field of the number of rounds. Each group gets attributed a number of rounds"""
-    last_round = models.IntegerField()
+    # """Field of the number of rounds. Each group gets attributed a number of rounds"""
+    # last_round = models.IntegerField()
+    pass
 
 
 class Player(BasePlayer):
