+++ introduction_cross/templates/introduction_cross/Instructions2.html
@@ -30,7 +30,7 @@
                     You <b>pay {{ Constants.c_high }} pts</b> for Participant <font color="blue">2</font> to <b>receive {{ Constants.b_high }} pts</b>.
                 </li>
                 <li>
-                    You <b>pay 0 pts</b>for Participant <font color="blue">2</font> to <b>receive 0 pts</b>.
+                    You <b>pay 0 pts</b> for Participant <font color="blue">2</font> to <b>receive 0 pts</b>.
                 </li>
             </ul>
             </div>
